# Speed up expected errors calculations using SSE2 instructions