--- a/src/cutadapt/expected_errors.h
+++ b/src/cutadapt/expected_errors.h
@@ -108,47 +108,41 @@ expected_errors_from_phreds(const uint8_t *phreds, size_t phreds_length, uint8_t
     uint8_t max_phred = 126 - base;
     #ifdef __SSE2__ 
     const uint8_t *vec_end_ptr = end_ptr - sizeof(__m128i);
-    __m128i phred_array;
-    __m128i illegal_phreds;
-    uint8_t phred_store[16];
     __m128 accumulator = _mm_set1_ps(0.0);
-    __m128 loader;
     while (cursor < vec_end_ptr) {
-        phred_array = _mm_loadu_si128((__m128i *)cursor);
-        illegal_phreds = _mm_cmpgt_epi8(phred_array, _mm_set1_epi8(max_phred));
+        __m128i phred_array = _mm_loadu_si128((__m128i *)cursor);
+        __m128i illegal_phreds = _mm_cmpgt_epi8(phred_array, _mm_set1_epi8(max_phred));
         illegal_phreds = _mm_or_si128(
             illegal_phreds, _mm_cmplt_epi8(phred_array, _mm_set1_epi8(base)));
         if (_mm_movemask_epi8(illegal_phreds)) {
             return -1.0;
         }
-        phred_array = _mm_sub_epi8(phred_array, _mm_set1_epi8(base));
-        _mm_store_si128((__m128i *)phred_store, phred_array);
-        loader = _mm_set_ps(
-            SCORE_TO_ERROR_RATE[phred_store[0]],
-            SCORE_TO_ERROR_RATE[phred_store[1]],
-            SCORE_TO_ERROR_RATE[phred_store[2]],
-            SCORE_TO_ERROR_RATE[phred_store[3]]
+        __m128 loader = _mm_set_ps(
+            SCORE_TO_ERROR_RATE[cursor[0] - base],
+            SCORE_TO_ERROR_RATE[cursor[1] - base],
+            SCORE_TO_ERROR_RATE[cursor[2] - base],
+            SCORE_TO_ERROR_RATE[cursor[3] - base]
         );
         accumulator = _mm_add_ps(accumulator, loader);
         loader = _mm_set_ps(
-            SCORE_TO_ERROR_RATE[phred_store[4]],
-            SCORE_TO_ERROR_RATE[phred_store[5]],
-            SCORE_TO_ERROR_RATE[phred_store[6]],
-            SCORE_TO_ERROR_RATE[phred_store[7]]
+            SCORE_TO_ERROR_RATE[cursor[4] - base],
+            SCORE_TO_ERROR_RATE[cursor[5] - base],
+            SCORE_TO_ERROR_RATE[cursor[6] - base],
+            SCORE_TO_ERROR_RATE[cursor[7] - base]
         );
         accumulator = _mm_add_ps(accumulator, loader);
         loader = _mm_set_ps(
-            SCORE_TO_ERROR_RATE[phred_store[8]],
-            SCORE_TO_ERROR_RATE[phred_store[9]],
-            SCORE_TO_ERROR_RATE[phred_store[10]],
-            SCORE_TO_ERROR_RATE[phred_store[11]]
+            SCORE_TO_ERROR_RATE[cursor[8] - base],
+            SCORE_TO_ERROR_RATE[cursor[9] - base],
+            SCORE_TO_ERROR_RATE[cursor[10] - base],
+            SCORE_TO_ERROR_RATE[cursor[11] - base]
         );
         accumulator = _mm_add_ps(accumulator, loader);
         loader = _mm_set_ps(
-            SCORE_TO_ERROR_RATE[phred_store[12]],
-            SCORE_TO_ERROR_RATE[phred_store[13]],
-            SCORE_TO_ERROR_RATE[phred_store[14]],
-            SCORE_TO_ERROR_RATE[phred_store[15]]
+            SCORE_TO_ERROR_RATE[cursor[12] - base],
+            SCORE_TO_ERROR_RATE[cursor[13] - base],
+            SCORE_TO_ERROR_RATE[cursor[14] - base],
+            SCORE_TO_ERROR_RATE[cursor[15] - base]
         );
         accumulator = _mm_add_ps(accumulator, loader);
         cursor += sizeof(__m128i);