For illumina this does not matter too much, for long reads however it is quite the difference.
Baseline
```
Benchmark 1: cutadapt ~/test/nanopore_500000reads.fastq -o /dev/null
  Time (mean ± σ):      4.040 s ±  0.019 s    [User: 3.020 s, System: 1.017 s]
  Range (min … max):    4.024 s …  4.072 s    5 runs
```
Calculating average error rate before
```
Benchmark 1: cutadapt ~/test/nanopore_500000reads.fastq -o /dev/null --max-aer 0.1
  Time (mean ± σ):      8.124 s ±  0.179 s    [User: 7.289 s, System: 0.814 s]
  Range (min … max):    7.989 s …  8.434 s    5 runs
```
After:
```
Benchmark 1: cutadapt ~/test/nanopore_500000reads.fastq -o /dev/null --max-aer 0.1
  Time (mean ± σ):      5.132 s ±  0.342 s    [User: 4.257 s, System: 0.858 s]
  Range (min … max):    4.834 s …  5.561 s    5 runs
```

Illumina also benefits slightly, but that is more in the 20% range rather than the 3x improvement range as shown here. This was build on #725. I will rebase on main once it is merged.

EDIT: I forget to mention, this uses floats instead for better speed. There is a slight precision loss, but this is quite insignificant given that phred scores have only two significant digits anyway. I could not find any filtering differences in the files I use for testing.