@marcelm 2023-09-06
This is a great speedup, very nice. I got confused because this PR presents itself as an alternative to #725, so I guess I made some review comments in the wrong place.