@rhpvorderman 2023-09-06
> This is a great speedup, very nice. I got confused because this PR presents itself as an alternative to https://github.com/marcelm/cutadapt/pull/725, so I guess I made some review comments in the wrong place.

Sure I will go update #725. I based this code on #725 so I could do some meaningful benchmarking. Sorry for creating this mess. Once #725 is merged the diff should look much cleaner.