@rhpvorderman 2023-09-06
Done. Rebased on main. I did a quick test to see if aligned loads makes a difference. It does not. So that saves some extra code.

I also did an avx2 solution over the weekend. According to the steam hardware survey, 90% of the users has avx2. I think in HPC systems it is actually closer to 100%. It will make compiling and building the project more difficult though. By contrast, the SSE2 solution just requires the `__SSE2__` build def which is automatically set to true by clang and gcc on any x86_64 machine. The speedup of avx2 over sse2 is another 2x but given the very small time that it takes generally not worth the hassle. 

https://github.com/rhpvorderman/cutadapt/blob/avx2expected_errors/src/cutadapt/expected_errors.h

In a few years from now, say 2027. I think it will be safe to say that x86_64 == AVX2 support and we can go truly wild. AVX2 is a supercool and very useful set.