@marcelm 2023-09-06
Nice, thanks again, will merge.

Regarding AVX2, we [had to disable it in strobealign](https://github.com/ksahlin/strobealign/issues/34). It makes strobealign about 10% faster overall, so that would be nice to have, but availability of ~91% is apparently not high enough. (In strobealign’s case, however, users can just recompile with the approriate flags and get the speedup.)