@rhpvorderman 2023-09-08
>  Regarding AVX2, we https://github.com/ksahlin/strobealign/issues/34. It makes strobealign about 10% faster overall, so that would be nice to have, but availability of ~91% is apparently not high enough

Yes, I think we have some execution nodes on our HPC also used that particular CPU (I think they are retired now). The solution for this is to use dynamic dispatch. You compile your function multiple times for different CPU's and make a selector function that updates its own pointer with the correct function when called (you only pay the overhead for choosing the function once). 
As you can  imagine, this is not as straightforward to implement and build. But it is an option when the extra performance is absolutely required. ISA-L does this for its functions and I am quite happy as this improves gzip compression/decompression by a lot. And we get to use that. Unfortunately it has quite a complex build system and it requires quite some expertise to set that up.

That is why I prefer using SSE2 so I can simply use the builtin `__SSE2__` setting that clang and gcc automatically activate on x86_64. Only MSVC does not do this, but that can be remedied quite easily. I think using `__SSE41__` and always compiling with `--msse4.1` would run practically on any x86_64 cpu. Minimap2 does that for instance. AVX I think is also quite ubiquitous by this point. Unfortunately AVX has barely any stuff that works on 8-bit integers, which is exactly what is required a lot in bionformatics. So it is not as interesting to enable that by default. 

I do love the SSE41 [_mm_blendv_epi8](https://www.intel.com/content/www/us/en/docs/intrinsics-guide/index.html#text=_mm_blendv&ig_expand=505) instruction. Combined with the comparison operations you can get a lot of work done. I am thinking of doing some changes to sequali that allow the algorithms to be vectorized a lot more. This blend function can be quite an interesting alternative to a lookup table for nucleotides which I use in sequali. Since I can do a _mm_cmpeq_epi8 for A,C,G,T,N and a _mm_blendv_epi8 to set the corresponding integer index (0 to 4) in the resulting array. This requires 10 operations. That should in theory be faster than 16 memory lookups (although they should all be on the same cache line, so this should be tested).