@marcelm 2023-09-11
> As you can imagine, this is not as straightforward to implement and build.

I looked into runtime dispatch a while ago, but decided because of the build system complexity that my time would be better spent on other methods of improving runtime and/or accuracy.

(BTW, runtime dispatch is a lot easier in Rust with just a couple of lines, see for example https://doc.rust-lang.org/core/arch/index.html#dynamic-cpu-feature-detection or https://alexheretic.github.io/posts/auto-avx2/#inlining .)

Have you heard that GCC and LLVM have recently gotten standardized [microarchitecture levels](https://www.phoronix.com/news/GCC-11-x86-64-Feature-Levels)? This adds support for `-march=x86-64-v2`, `-march=x86-64-v3` etc. I find it helpful to think about CPU support in terms of which level one can aim for.