@rhpvorderman 2023-09-11
> I looked into runtime dispatch a while ago, but decided because of the build system complexity that my time would be better spent on other methods of improving runtime and/or accuracy.

Yes, I feel this is a very last effort approach when performance is absolutely key. When working with 8-bit integers the theoretical speedup of sse2 can already by 16x. Going to 32x is nice, but basically feels from going from almost no time, to half of almost no time. Not as compelling, giving the required effort.

> (BTW, runtime dispatch is a lot easier in Rust with just a couple of lines, see for example https://doc.rust-lang.org/core/arch/index.html#dynamic-cpu-feature-detection or https://alexheretic.github.io/posts/auto-avx2/#inlining .)

Interesting. Supposedly it is also quite doable using some GCC specific commands, but then you are committing to GCC only, so that is not ideal. As for Rust, I am so happy with Python doing all the non-performance critical stuff, and C (or Cython) works so well with python that I have little reason to look into it. Maybe some time when I can contribute to a major project written in Rust?

> Have you heard that GCC and LLVM have recently gotten standardized [microarchitecture levels](https://www.phoronix.com/news/GCC-11-x86-64-Feature-Levels)? This adds support for -march=x86-64-v2, -march=x86-64-v3 etc. I find it helpful to think about CPU support in terms of which level one can aim for.

Thanks! That is massively helpful.